+++ src/server/pipes/src/chan.rs
@@ -223,7 +223,7 @@ impl Channel {
 
         // commits are done here, because they don't get new data
         if commit > 0 {
-            return reply_vmsg!(is, Code::Success as u32, state.rbuf.size());
+            return reply_vmsg!(is, Code::Success, state.rbuf.size());
         }
 
         // if there are already queued read requests, just append this request
@@ -250,13 +250,13 @@ impl Channel {
                 amount,
                 pos
             );
-            reply_vmsg!(is, Code::Success as u32, pos, amount)
+            reply_vmsg!(is, Code::Success, pos, amount)
         }
         else {
             // nothing to read; if there is no writer left, report EOF
             if state.flags().contains(Flags::WRITE_EOF) {
                 log!(crate::LOG_DEF, "[{}] pipes::read(): EOF", self.id);
-                reply_vmsg!(is, Code::Success as u32, 0usize, 0usize)
+                reply_vmsg!(is, Code::Success, 0usize, 0usize)
             }
             else {
                 // if we promised the client that input would be available, report WouldBlock
@@ -308,7 +308,7 @@ impl Channel {
 
         // commits are done here, because they don't get new data
         if commit > 0 {
-            return reply_vmsg!(is, Code::Success as u32, state.rbuf.size());
+            return reply_vmsg!(is, Code::Success, state.rbuf.size());
         }
 
         // if there are already queued write requests, just append this request
@@ -332,7 +332,7 @@ impl Channel {
                 amount,
                 pos
             );
-            reply_vmsg!(is, Code::Success as u32, pos, amount)
+            reply_vmsg!(is, Code::Success, pos, amount)
         }
         else {
             // if we promised the client that input would be available, report WouldBlock
